requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.11
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-timezone-field==4.1.2
gitdb==4.0.7
GitPython==3.1.17
kombu==5.1.0
lxml==4.6.3
Pillow==8.1.2
prompt-toolkit==3.0.18
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
psycopg2-binary~=2.8.6
django-celery-beat~=2.2.0
celery~=5.1.0
redis~=3.5.3
